+++ components/brave_ads/core/internal/creatives/promoted_content_ads/creative_promoted_content_ad_test_util.h
@@ -6,11 +6,13 @@
 #ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_CREATIVES_PROMOTED_CONTENT_ADS_CREATIVE_PROMOTED_CONTENT_AD_TEST_UTIL_H_
 #define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_CREATIVES_PROMOTED_CONTENT_ADS_CREATIVE_PROMOTED_CONTENT_AD_TEST_UTIL_H_
 
+#include <cstddef>
+
 #include "brave/components/brave_ads/core/internal/creatives/promoted_content_ads/creative_promoted_content_ad_info.h"
 
 namespace brave_ads::test {
 
-CreativePromotedContentAdList BuildCreativePromotedContentAds(int count);
+CreativePromotedContentAdList BuildCreativePromotedContentAds(size_t count);
 CreativePromotedContentAdInfo BuildCreativePromotedContentAd(
     bool should_generate_random_uuids);
